more error handling, email fields are unique now, centralized test file in /src
